fixed some type error flagged by compiler

+++ Source/Objects/Hardware/Vme/GretinaTrigger/ORGretinaTriggerController.m
@@ -340,7 +340,7 @@
 
 - (void) linkLockedRegChanged:(NSNotification*)aNote
 {
-    int value = ~[model linkLockedReg];
+    NSInteger value = ~[model linkLockedReg];
 	short i;
 	for(i=0;i<[linkLockedMatrix numberOfColumns];i++){
 		[[linkLockedMatrix cellWithTag:i] setIntValue:(value & 1L<<i)];
@@ -349,7 +349,7 @@
 
 - (void) linkLruCrlRegChanged:(NSNotification*)aNote
 {
-    int value = [model linkLruCrlReg];
+    NSInteger value = [model linkLruCrlReg];
 	short i;
 	for(i=0;i<[linkLruCrlMatrix numberOfColumns];i++){
 		[[linkLruCrlMatrix cellWithTag:i] setIntValue:(value & 1L<<i)];
@@ -358,7 +358,7 @@
 
 - (void) miscCtl1RegChanged:(NSNotification*)aNote
 {
-    int value = [model miscCtl1Reg];
+    NSInteger value = [model miscCtl1Reg];
 	short i;
 	for(i=0;i<[miscCtl1Matrix numberOfColumns];i++){
 		[[miscCtl1Matrix cellWithTag:i] setIntValue:(value & 1L<<i)];
@@ -372,7 +372,7 @@
 
 - (void) inputLinkMaskChanged:(NSNotification*)aNote
 {
-    int value = ~[model inputLinkMask];
+    NSInteger value = ~[model inputLinkMask];
 	short i;
 	for(i=0;i<[inputLinkMaskMatrix numberOfColumns];i++){
 		[[inputLinkMaskMatrix cellWithTag:i] setIntValue:(value & 1L<<i)];
@@ -381,7 +381,7 @@
 
 - (void) serDesTPowerMaskChanged:(NSNotification*)aNote
 {
-    int value = [model serdesTPowerMask];
+    NSInteger value = [model serdesTPowerMask];
 	short i;
 	for(i=0;i<[serDesTPowerMasMatrix numberOfColumns];i++){
 		[[serDesTPowerMasMatrix cellWithTag:i] setIntValue:(value & 1L<<i)];
@@ -390,7 +390,7 @@
 
 - (void) lvdsPreemphasisCtlChanged:(NSNotification*)aNote
 {
-    int value = [model lvdsPreemphasisCtlMask];
+    NSInteger value = [model lvdsPreemphasisCtlMask];
 	short i;
 	for(i=0;i<[lvdsPreemphasisCtlMatrix numberOfColumns];i++){
 		[[lvdsPreemphasisCtlMatrix cellWithTag:i] setIntValue:(value & 1L<<i)];
@@ -399,7 +399,7 @@
 
 - (void) serDesRPowerMaskChanged:(NSNotification*)aNote
 {
-    int value = [model serdesRPowerMask];
+    NSInteger value = [model serdesRPowerMask];
 	short i;
 	for(i=0;i<[serDesRPowerMasMatrix numberOfColumns];i++){
 		[[serDesRPowerMasMatrix cellWithTag:i] setIntValue:(value & 1L<<i)];
